[#66] Enable client to run independently of iRODS server
This change enables the iRODS C++ REST client to run independently of an
iRODS server on the local host by removing dependencies on entitites
which are included in the iRODS server packages and configuration in a
set up iRODS server.

The following changes were made to support this effort:

Fix packaging for Centos 7. This addresses issue #132.

---

Remove unused function configure_logger

init_logger seems to be the replacement and is being used throughout the
files with main() functions to configure log levels for each REST
service.

---

Match installation paths with project name (irods_client_rest_cpp).
Additionally, the postinst script has been removed from the built
packages as it will no longer functions on hosts which do not have an
irods Linux user or irods Linux group.

---

Remove iRODS dependencies in init.d script

The script used by init.d to manage the REST API endpoints used several
things which required an iRODS installation on the same computer in
order to run.

A script to determine the OS flavor is packaged with the iRODS server
and was being used in this script as well because this script is based
heavily on the one used for the iRODS server. The contents of find_os
were dropped into this script and captured in a function.

The starting and stopping of the API endpoints is performed by the iRODS
service account. The REST client does not need to have a service account
to run, so these calls were changed to just be run as the current user.

---

Rework configuration class

Moved configuration class to its own compilation unit with the intention
to grow a "common" shared object against which the REST services can
link.

Instead of tying the configuration class to a particular REST service
name as in the past, this implements a series of free functions which
provide a read-only interface to the configuration which is now a static
nlohmann::json object.

The configuration file is loaded once at program start and survives for
the duration of the program. This means that configuration changes
require a REST service restart.

This also updates the JSON configuration template file to include the
irods_client_environment and rest_api keys.

The JSON configuration file has now expanded to include an
irods_client_environment to hold information about the iRODS client
and server. A rest_api section has been added to hold the REST service
configuration options as well as a jwt_signing_key to serve as the
signing key for JWTs instead of using the zone_key found in the iRODS
server_config.json file. This enables the REST client to stand entirely
independently of other clients as well as the iRODS server itself.

The iRODS client environment is now held entirely in the JSON
configuration file and we do not need to look at the irods_environment
file anymore.

---

Save rodsadmin password in .irodsA file

Since native authentication is the only supported authentication method
for the REST client, we need to make sure that the obfuscated password
file is created beforehand in order to satisfy the password entry
requirement. This is normally handled by iinit or a prompt provided by
the native authentication plugin in the tty. Seeing as there is no tty
with which to interact, the rodsadmin proxy user needs to be
authenticated beforehand.

This change checks for the obfuscated password file using the standard
client utility from the iRODS runtime package - obfGetPw - and uses the
standard client utility from the iRODS runtime package for saving the
obfuscated password - obfSavePw - by using the rodsadmin_password entry
in the irods_client_environment object of the client JSON configuration
file.

---

Bump version number to 0.9.2 and update the README appropriately.

---

Removed header comments in *Main*.cpp files about the code being
generated because it's not true anymore.

---

Surrounded main() function bodies with a try-catch that handles all
exceptions by returning a code of 1. In the future, more elegant error
handling can be implemented.