[irods/irods#6527] Update clang-format and add clang-tidy
The clang-format and clang-tidy files were copied directly from the main
irods/irods repository and can be compared/contrasted with this:

https://github.com/irods/irods/blob/6b8d2dc56af72c81a568e22df78eac7b9cd745ef/.clang-format
https://github.com/irods/irods/blob/6b8d2dc56af72c81a568e22df78eac7b9cd745ef/.clang-tidy

A Github Actions workflow was added for clang-tidy based on the workflow
from the main irods/irods repository. Some installed packages were
adjusted to fit the needs of the iRODS REST C++ client.