[#143] Add Compose project for running tests
This adds a docker-compose project which acts very similarly to the
docker/runner project, except that an iRODS server is installed and set
up on the container running the REST client service. Instead of keeping
the service alive via forever-sleep, the test suite is run using
irods.test.settings.ICAT_HOSTNAME as the name of the iRODS server host
and irods.test.settings.HOSTNAME_1 as the name of the nginx
reverse-proxy service.

Also, adds all .deb files to .gitignore.