[#134] Add test for /list with stat=1 on collection
Also changed the irods_rest.py interface for list to use boolean and
integer values with default values. The URL must specify all options in
order for the /list endpoint to understand it, but the python interface
used by the tests can now specify only what is not going to use the
default values.